backend: add support for offset and limit in GET request
In the URI of the GET requests for all the users add the following
parameters $offset and $limit, like the following URI

GET /cytech/user?offset=0&limit=3

where the $offset is the number of rows to skip before starting to
return rows from the query and the $limit is the maximum number of rows
to return. In the example above, fetch the first 3 rows from the
database.